+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/IrOnlyClassesBoxTestGenerated.java
@@ -3555,6 +3555,18 @@ public class IrOnlyClassesBoxTestGenerated extends AbstractIrOnlyClassesBoxTest 
     }
 
     @Test
+    @TestMetadata("MethodAndFieldInjectionKeepDependencyOrder.kt")
+    public void testMethodAndFieldInjectionKeepDependencyOrder() {
+      run("MethodAndFieldInjectionKeepDependencyOrder.kt");
+    }
+
+    @Test
+    @TestMetadata("MethodBeforeFieldInjectionWithoutInterop.kt")
+    public void testMethodBeforeFieldInjectionWithoutInterop() {
+      run("MethodBeforeFieldInjectionWithoutInterop.kt");
+    }
+
+    @Test
     @TestMetadata("ScopedInjectIntoBaseMember.kt")
     public void testScopedInjectIntoBaseMember() {
       run("ScopedInjectIntoBaseMember.kt");