+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/JsFastInitBoxTestGenerated.java
@@ -2619,6 +2619,18 @@ public class JsFastInitBoxTestGenerated extends AbstractJsFastInitBoxTest {
     }
 
     @Test
+    @TestMetadata("MethodAndFieldInjectionKeepDependencyOrder.kt")
+    public void testMethodAndFieldInjectionKeepDependencyOrder() {
+      run("MethodAndFieldInjectionKeepDependencyOrder.kt");
+    }
+
+    @Test
+    @TestMetadata("MethodBeforeFieldInjectionWithoutInterop.kt")
+    public void testMethodBeforeFieldInjectionWithoutInterop() {
+      run("MethodBeforeFieldInjectionWithoutInterop.kt");
+    }
+
+    @Test
     @TestMetadata("ScopedInjectIntoBaseMember.kt")
     public void testScopedInjectIntoBaseMember() {
       run("ScopedInjectIntoBaseMember.kt");